Update the request and response after pipelines finish processing
This change became required after completing
cbc843g in https://github.com/refinery29/piston/pull/121cbc843f. Which
changes Piston to work in an immutable manner.

The reason why is that the _original_ object is changed during
the pipeline processing. But when working immutably, and a change is made,
then the original object is no longer used, instead replaced with a new
object, which contains all of the state of the original, yet with the
required changes made.

Therefore, we have to be more pedantic in how the object is used
throughout the application call stack. Specifically, the request, and
response objects have to be returned after the respective pipelines have
finished processing, so that Piston can be updated with them. Otherwise
the changes made throughout the pipeline won't be persisted.